# public/openapi.json
{
    "openapi":  "3.0.0",
    "info":  {
                 "title":  "Empresa de Transportes El Cumbe - API RESTful",
                 "version":  "1.0.0",
                 "description":  "EspecificaciÃ³n oficial de servicios RESTful y API MÃ³vil para la plataforma de transporte interprovincial El Cumbe. Incluye servicios de autenticaciÃ³n JWT, bÃºsqueda de salidas, reserva de asientos, pasarela de pagos Culqi y gestiÃ³n de encomiendas."
             },
    "servers":  [
                    {
                        "url":  "http://localhost:3000",
                        "description":  "Servidor Local de Desarrollo"
                    }
                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT",
                                                                  "description":  "Ingrese el token JWT obtenido en la API de Login MÃ³vil (/api/movil/login)"
                                                              }
                                           },
...
                   },
    "paths":  {
...
                  "/api/movil/conductor/alertas":  {
                                                       "get":  {
                                                                   "tags":  [
                                                                                "Conductor"
                                                                            ],
                                                                   "summary":  "Listar alertas propias",
                                                                   "security":  [
                                                                                    {
                                                                                        "bearerAuth":  [

                                                                                                       ]
                                                                                    }
                                                                                ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Operación exitosa"
                                                                                             },
                                                                                     "400":  {
                                                                                                 "description":  "Solicitud inválida"
                                                                                             },
                                                                                     "401":  {
                                                                                                 "description":  "Token ausente o inválido"
                                                                                             },
                                                                                     "403":  {
                                                                                                 "description":  "Rol o recurso no autorizado"
                                                                                             },
                                                                                     "500":  {
                                                                                                 "description":  "Error interno"
                                                                                             }
                                                                                 }
                                                               },
                                                       "post":  {
                                                                    "tags":  [
                                                                                 "Conductor"
                                                                             ],
                                                                    "summary":  "Marcar una alerta propia como leída",
                                                                    "security":  [
                                                                                     {
                                                                                         "bearerAuth":  [

                                                                                                        ]
                                                                                     }
                                                                                 ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Operación exitosa"
                                                                                              },
                                                                                      "400":  {
                                                                                                  "description":  "Solicitud inválida"
                                                                                              },
                                                                                      "401":  {
                                                                                                  "description":  "Token ausente o inválido"
                                                                                              },
                                                                                      "403":  {
                                                                                                  "description":  "Rol o recurso no autorizado"
                                                                                              },
                                                                                      "500":  {
                                                                                                  "description":  "Error interno"
                                                                                              }
                                                                                  },
                                                                    "requestBody":  {
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "type":  "object"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    },
                                                                                        "required":  true
                                                                                    }
                                                                }
                                                   },
                  "/api/movil/conductor/bitacora":  {
                                                        "post":  {
                                                                     "tags":  [
                                                                                  "Conductor"
                                                                              ],
                                                                     "summary":  "Registrar una ocurrencia en un viaje propio",
                                                                     "security":  [
                                                                                      {
                                                                                          "bearerAuth":  [

                                                                                                         ]
                                                                                      }
                                                                                  ],
                                                                     "responses":  {
                                                                                       "200":  {
                                                                                                   "description":  "Operación exitosa"
                                                                                               },
                                                                                       "400":  {
                                                                                                   "description":  "Solicitud inválida"
                                                                                               },
                                                                                       "401":  {
                                                                                                   "description":  "Token ausente o inválido"
                                                                                               },
                                                                                       "403":  {
                                                                                                   "description":  "Rol o recurso no autorizado"
                                                                                               },
                                                                                       "500":  {
                                                                                                   "description":  "Error interno"
                                                                                               }
                                                                                   },
                                                                     "requestBody":  {
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "type":  "object"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     },
                                                                                         "required":  true
                                                                                     }
                                                                 },
                                                        "put":  {
                                                                    "tags":  [
                                                                                 "Conductor"
                                                                             ],
                                                                    "summary":  "Resolver una ocurrencia propia",
                                                                    "security":  [
                                                                                     {
                                                                                         "bearerAuth":  [

                                                                                                        ]
                                                                                     }
                                                                                 ],
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "Operación exitosa"
                                                                                              },
                                                                                      "400":  {
                                                                                                  "description":  "Solicitud inválida"
                                                                                              },
                                                                                      "401":  {
                                                                                                  "description":  "Token ausente o inválido"
                                                                                              },
                                                                                      "403":  {
                                                                                                  "description":  "Rol o recurso no autorizado"
                                                                                              },
                                                                                      "500":  {
                                                                                                  "description":  "Error interno"
                                                                                              }
                                                                                  },
                                                                    "requestBody":  {
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "type":  "object"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    },
                                                                                        "required":  true
                                                                                    }
                                                                }
                                                    },
...
              }
}